requirements.txt
python-dotenv
requests
selenium
selenium-wire
google-auth
google-auth-oauthlib
google-api-python-client
